refactor(analysis-types): split duplication DTOs
